test: run configured suites from repository root
Move the core version assertions into the conventionally collected tests suite and run pytest once from the repository root so tests/, common/, and dblib/ are all covered.

Tested:
- myenv/bin/pytest --collect-only -q (332 collected)
- myenv/bin/pytest -q tests/test_core_version.py (1 passed)
- myenv/bin/pytest (332 passed)
